README now explains how to run and test the prototype
I've written up the full process on how to install and run our project, then test it.  This includes diagrams and links to various resources used.  PLEASE NOTE: As I had never manually connected a physical device to run our project, I don't know the exact flow.  In other words, that part of the readme may need to be updated.